renoir: overlay: switch statusbar height from dp to px
Looks like there is a bug on the framework side when converting
dp to px, especially when a video playback app is in fullscreen
landscape mode. When pulling down once to reveal the statusbar,
the content of the statusbar suddenly shifts down a little, possibly
because a new statusbar height in px is calculated with an incorrect
screen density.

Temporarily fix this by setting statusbar height in px directly.

Conversion formula [1]:
height in pixels = 30dp * 560 screen density / 160 = 105px

[1] https://material.io/design/layout/pixel-density.html#pixel-density-on-android

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>